# api-spec/src/main/resources/static/api.yml
...
swagger: '2.0'
info:
  title: Name of the App API
  description: Use our services using this API
  version: "1.0.0"
# the domain of the service
host: api.app.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /rules:
    post:
      summary: Create rule
      description: |
        This method creates a new rule.
      parameters:
        - name: rule_id
          in: query
          description: Unique identifier representing a specific rule.
          required: true
          type: string
        - name: near_enter
          in: query
          description: Near from the geofence, entering.
          required: true
          type: boolean
        - name: near_leave
          in: query
          description: Near from the geofence, leaving.
          required: true
          type: boolean
        - name: enter
          in: query
          description: Entering the geofence.
          required: true
          type: boolean
        - name: leave
          in: query
          description: Leaving the geofence.
          required: true
          type: boolean
        - name: inside
          in: query
          description: Inside of the geofence.
          required: true
          type: boolean
        - name: distance
          in: query
          description: Distance from the geofence.
          required: true
          type: integer
        - name: time
          in: query
          description: Time of waiting.
          required: true
          type: integer
      tags:
        - Rules
      responses:
        201:
          description: Rule created
          schema:
            $ref: '#/definitions/Rule'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get an array of rules
      description: |
        This method returns an array of rules from a certain user.
      parameters:
        - name: user
          in: query
          description: Owner of the rules to be returned.
          required: true
          type: string
      tags:
        - Rules
      responses:
        200:
          description: Rules from the user
          schema:
            type: array
            items:
              $ref: '#/definitions/Rule'
        404:
          description: Rule not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /rules/{rule_id}:
    put:
      summary: Modify rule
      description: |
        This method modifies the data of a previously created rule.
      parameters:
        - name: rule_id
          in: path
          description: Unique identifier representing a specific rule.
          required: true
          type: string
        - name: near_enter
          in: query
          description: Near from the geofence, entering.
          required: false
          type: boolean
        - name: near_leave
          in: query
          description: Near from the geofence, leaving.
          required: false
          type: boolean
        - name: enter
          in: query
          description: Entering the geofence.
          required: false
          type: boolean
        - name: leave
          in: query
          description: Leaving the geofence.
          required: false
          type: boolean
        - name: inside
          in: query
          description: Inside of the geofence.
          required: false
          type: boolean
        - name: distance
          in: query
          description: Distance from the geofence.
          required: false
          type: integer
        - name: time
          in: query
          description: Time of waiting.
          required: false
          type: integer
      tags:
        - Rules
      responses:
        202:
          description: Rule's data modified
          schema:
            $ref: '#/definitions/Rule'
        304:
          description: Rule's data not modified
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Rule not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete rule
      description: |
        This method deletes the data of a previously created rule.
      parameters:
        - name: rule_id
          in: path
          description: Unique identifier representing a specific rule.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the rule.
          required: true
          type: string
      tags:
        - Rules
      responses:
        200:
          description: Rule deleted
          schema:
            $ref: '#/definitions/Rule'
        404:
          description: Rule not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get rule by rule_id
      description: |
        This method returns a rule by rule_id.
      parameters:
        - name: rule_id
          in: path
          description: Unique identifier representing a specific rule.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the notification.
          required: true
          type: string
      tags:
        - Rules
      responses:
        200:
          description: Rule with rule_id requested
          schema:
            $ref: '#/definitions/Rule'
        404:
          description: Rule not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Rule:
    type: object
    properties:
      rule_id:
        type: string
        description: Unique identifier representing a specific rule.
      near_enter:
        type: boolean
        description: Near from the geofence, entering.
      near_leave:
        type: boolean
        description: Near from the geofence, leaving.
      enter:
        type: boolean
        description: Entering the geofence.
      leave:
        type: boolean
        description: Leaving the geofence.
      inside:
        type: boolean
        description: Inside of the geofence.
      distance:
        type: integer
        description: Distance from the geofence.
      time:
        type: integer
        description: Time of waiting.
  Error:
    type: object
    properties:
      title:
        type: string
      body:
        type: string